--- a/vng/vector/primitive.go
+++ b/vng/vector/primitive.go
@@ -183,14 +183,6 @@ type PrimitiveReader struct {
 	it  zcode.Iter
 }
 
-// TODO Remove these before merging
-func (p *PrimitiveReader) Segmap() []Segment {
-	return p.segmap
-}
-func (p *PrimitiveReader) Reader() io.ReaderAt {
-	return p.reader
-}
-
 func NewPrimitiveReader(primitive *Primitive, reader io.ReaderAt) *PrimitiveReader {
 	return &PrimitiveReader{
 		Typ:    primitive.Typ,
@@ -242,14 +234,6 @@ type DictReader struct {
 	dict      []DictEntry
 	selectors []byte
 	off       int
-}
-
-// TODO Remove these before merging
-func (p *DictReader) Segmap() []Segment {
-	return p.segmap
-}
-func (p *DictReader) Reader() io.ReaderAt {
-	return p.reader
 }
 
 func NewDictReader(primitive *Primitive, reader io.ReaderAt) *DictReader {