… vector/vector_test.go
… vng/vector/spiller.go
… vng/ztests/const.yaml