* Read directly from segmap rather than using `ReadBytes`. 
* Include a count in each segment (this is a breaking change to vng). Use this to pre-size slices.
* Reuse a single buffer across all internal read calls.
* Avoid unnecessary copies for string/bytes vectors.
* Correctly handle dict vectors, now that they're enabled.
* Add a benchmark tracking vng read performance and comparing it to zng and raw varints.

On wrccdc_mono_10m.vng this cuts vector.Read time from 270ms -> 140ms for id.resp_p and 470ms -> 180ms for uid.

Materializer performance is still terrible. I had some success improving it by only exposing a Puller instead of a zio.Reader, but the changes ended up being quite invasive elsewhere so I backed it out. It's probably worth trying again later once we only have one set of code for reading vng.